core: move DistanceRange{Map,Set} to osrd-mp
Needed because the train simulation depends on it
