+++ core/osrd-mp/src/commonMain/kotlin/fr/sncf/osrd/utils/DistanceRangeSet.kt
